LottoService의 로직을 LottoResult로 옮기기
