+++ src/main/java/LottoApplication.java
@@ -5,7 +5,6 @@ import domain.lotto.LottoResult;
 import domain.lotto.LottoStore;
 import domain.lotto.LottoTicket;
 import domain.lotto.dto.StatistsDto;
-import domain.lotto.service.LottoService;
 import java.util.List;
 import view.InputView;
 import view.OutputView;
@@ -26,8 +25,7 @@ public class LottoApplication {
 
         final LottoResult lottoResult = new LottoResult(winningTicket, bonusNumber);
 
-        final LottoService lottoService = new LottoService();
-        StatistsDto statistsDto = lottoService.makeStatistics(money, lottoResult, lottoTickets);
+        StatistsDto statistsDto = lottoResult.makeStatistics(money, lottoTickets);
 
         OutputView.printStatistics(statistsDto);
     }
